# api/key-distribution-openapi.yaml
openapi: 3.0.0
info:
  description: Pay Super Key Products api
  version: 1.0.0
  title: Pay Super key distribution
  contact:
    email: support@protocol.one
...
servers:
  - description: Master test server
    url: https://p1payapi.tst.protocol.one
paths:
...
  /admin/api/v1/key-products/:key_product_id/platforms:
    get:
      tags:
        - product
        - platform
      summary: Get platforms with prices for specified key product
      parameters:
        - name: key_product_id
          in: path
          description: "Key Product ID"
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KeyPlatform'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - product
        - platform
      summary: Change prices for specified platform and key product
      parameters:
        - name: key_product_id
          in: path
          description: "Key Product ID"
          required: true
          schema:
            type: string
        - name: platform_id
          in: path
          description: "Platform id"
          required: true
          example: steam
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                platform:
                  $ref: '#/components/schemas/KeyPlatform'
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'
...
  /api/v1/key-products/:key_product_id:
    get:
      tags:
        - product
      summary: "Get product with platforms list and their prices"
      parameters:
        - name: key_product_id
          in: path
          description: "Key Product ID"
          required: true
          schema:
            type: string
        - name: country
          in: query
          description: ISO code of country for that price shoul be calculated
          required: false
          example: RUS
          schema:
            type: string
        - name: currency
          in: query
          description: Currency of price that should be calculated
          example: RUB
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyProductInfo'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal Server Error
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    KeyPlatform:
      type: object
      properties:
        id:
          type: string
          example: gog
        name:
          type: string
          example: Good old games.com
        prices:
          type: array
          items:
            $ref: '#/components/schemas/KeyProductPrice'
    KeyProductPrice:
      type: object
      properties:
        region:
          type: string
        amount:
          type: number
          format: float
          example: 29.99
        currency:
          type: string
          example: EUR
          enum:
            - EUR
            - USD
            - RUB
...
    KeyProductInfo:
      type: object
      properties:
        id:
          type: string
          description: id of product
        project_id:
          type: string
          description: Id of project
        name:
          type: string
          description: Localized name of product
        description:
          type: string
          description: Localized description of description
        long_description:
          type: string
          description: Localized long description of description
        images:
          type: array
          items:
            type: string
        platforms:
          type: array
          items:
            $ref: '#/components/schemas/KeyPlatformInfo'
    KeyPlatformInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: object
          properties:
            amount:
              type: number
            currency:
              type: string
              example: USD
            region:
              type: string
              example: USD
            is_fallback:
              type: boolean
              description: Flag indicates that price was calculated in default currency instead of specified
...